+++ projectroles/tests/test_views_taskflow.py
@@ -908,7 +908,7 @@ class TestRoleAssignmentOwnerTransferView(TestTaskflowBase):
         with self.login(self.user):
             self.client.post(
                 reverse(
-                    'projectroles:role_transfer_owner',
+                    'projectroles:role_owner_transfer',
                     kwargs={'project': self.project.sodar_uuid},
                 ),
                 data={