+++ CHANGELOG.rst
@@ -39,6 +39,7 @@ Changed
     - Allow updating local app settings on a ``TARGET`` site (#545)
     - Refactor project list filtering (#566)
     - Move project list javascript to ``project_list.js`` (#566)
+    - Rename owner role transfer URL pattern and timeline event (#590)
 - **Userprofile**
     - Cosmetic updates for user detail template (#600)